添加按下飞机跳过启动动画，将启动Activity修改成StartActivity
